Moved duplicated test code to a new file.
