+++ Cargo.toml
@@ -17,13 +17,13 @@ asynchronous = "~0.4.5"
 flate2 = "~0.2.12"
 maidsafe_utilities = "~0.1.5"
 memmap = "~0.2.3"
+rand = "~0.3.13"
 rustc-serialize = "~0.3.16"
 sodiumoxide = "~0.0.9"
 
 [dev-dependencies]
 cbor = "~0.3.16"
 docopt = "~0.6.78"
-rand = "~0.3.13"
 
 [[example]]
 name = "basic_encryptor"
